Try and match the standard when validating emails.
We ran into an issue where "e@e.e" passes client-side validation on
Chrome (and maybe others) but blows up in Formulary.  We were using a
gem that has a stricter validation than the standard.  We sort of
borrowed a regex from Stackoverflow, which is slightly worrying but it
passes the tests...